Load image from canvas extracted to a new file
