+++ src/components/LoadImgFile.jsx
@@ -1,13 +1,11 @@
 import React, { useState, useRef, useEffect } from 'react';
-import { fromJS } from 'immutable';
-import shortid from 'shortid';
 import styled, { css } from 'styled-components';
-import getTimeInterval from '../utils/intervals';
 import Picker from './Picker';
 import Button from './Button';
 import ValidationMessage from './ValidationMessage';
 import breakpoints from '../utils/breakpoints';
 import drawFileImageToCanvas from '../utils/ImageToCanvas';
+import generateFrames from '../utils/loadFromCanvas';
 
 const Container = styled.div`
   text-align: center;
@@ -117,7 +115,7 @@ const LoadImgFile = props => {
         canvasRef.current,
         () => setImageLoaded(true)
       );
-      if (validationData) {
+      if (validationData.errorType) {
         setImageLoaded(false);
         showValidationMessage({
           show: true,
@@ -133,62 +131,62 @@ const LoadImgFile = props => {
     }
   };
 
-  const getHeightIntervals = (imageHeight, numberOfFrames) => {
-    const heightPerFrame = Math.floor(imageHeight / numberOfFrames);
-    const intervals = [];
-    let top = 0;
-    let bottom = heightPerFrame;
-    for (let i = 0; i < numberOfFrames; i++) {
-      intervals.push({
-        top,
-        bottom,
-        timePercentage: getTimeInterval(i, numberOfFrames)
-      });
-      top += heightPerFrame;
-      bottom += heightPerFrame;
-    }
-    return intervals;
-  };
+  // const getHeightIntervals = (imageHeight, numberOfFrames) => {
+  //   const heightPerFrame = Math.floor(imageHeight / numberOfFrames);
+  //   const intervals = [];
+  //   let top = 0;
+  //   let bottom = heightPerFrame;
+  //   for (let i = 0; i < numberOfFrames; i++) {
+  //     intervals.push({
+  //       top,
+  //       bottom,
+  //       timePercentage: getTimeInterval(i, numberOfFrames)
+  //     });
+  //     top += heightPerFrame;
+  //     bottom += heightPerFrame;
+  //   }
+  //   return intervals;
+  // };
 
-  const generateFrames = (imageContext, numberOfFrames, pixSize = 1) => {
-    const { width, height } = imageContext.canvas;
-    const heightIntervals = getHeightIntervals(height, numberOfFrames);
-    const frameCollection = [];
+  // const generateFrames = (imageContext, numberOfFrames, pixSize = 1) => {
+  //   const { width, height } = imageContext.canvas;
+  //   const heightIntervals = getHeightIntervals(height, numberOfFrames);
+  //   const frameCollection = [];
 
-    heightIntervals.forEach(heightInterval => {
-      const pixelWidth = pixSize;
-      const pixelHeight = pixSize;
+  //   heightIntervals.forEach(heightInterval => {
+  //     const pixelWidth = pixSize;
+  //     const pixelHeight = pixSize;
 
-      console.log(heightInterval.top, heightInterval.bottom);
+  //     console.log(heightInterval.top, heightInterval.bottom);
 
-      const grid = [];
-      for (
-        let y = heightInterval.top;
-        y + pixelHeight <= heightInterval.bottom;
-        y += pixelWidth
-      ) {
-        for (let x = 0; x + pixelWidth <= width; x += pixelWidth) {
-          const currentPixel = imageContext.getImageData(
-            x,
-            y,
-            pixelWidth,
-            pixelHeight
-          ).data;
-          grid.push(
-            `rgba(${currentPixel[0]},${currentPixel[1]},${currentPixel[2]},${currentPixel[3]})`
-          );
-        }
-      }
+  //     const grid = [];
+  //     for (
+  //       let y = heightInterval.top;
+  //       y + pixelHeight <= heightInterval.bottom;
+  //       y += pixelWidth
+  //     ) {
+  //       for (let x = 0; x + pixelWidth <= width; x += pixelWidth) {
+  //         const currentPixel = imageContext.getImageData(
+  //           x,
+  //           y,
+  //           pixelWidth,
+  //           pixelHeight
+  //         ).data;
+  //         grid.push(
+  //           `rgba(${currentPixel[0]},${currentPixel[1]},${currentPixel[2]},${currentPixel[3]})`
+  //         );
+  //       }
+  //     }
 
-      frameCollection.push({
-        grid,
-        interval: heightInterval.timePercentage,
-        key: shortid.generate()
-      });
-    });
+  //     frameCollection.push({
+  //       grid,
+  //       interval: heightInterval.timePercentage,
+  //       key: shortid.generate()
+  //     });
+  //   });
 
-    return fromJS(frameCollection);
-  };
+  //   return fromJS(frameCollection);
+  // };
 
   const loadImgValidation = (context, size, frameAmount) => {
     const maxPixelsWidth = 100;
